fix(lens): The endpoint reaches past `connect` to dial the socket itself for readiness — the transport-reach volatility leaks across the receptacle
Factored the unix-socket connect/error race into a single supervisor-exported dialSocket() used by both endpoint.ts's readiness probe and ptyHost/connect.ts's dialer.

Agreed by the lowy ⇄ hickey lens debate (finding lowy-3, raised by lowy). Not pushed or merged.
